[OM-74542] Implement/Change the switch over functionality to use the newly created MySQL tables

See merge request turbonomic/xl!3231

+++ com.vmturbo.search.metadata/src/main/java/com/vmturbo/search/metadata/DbFieldDescriptor.java
@@ -1,7 +1,5 @@
 package com.vmturbo.search.metadata;
 
-import java.math.BigDecimal;
-
 import org.jooq.DataType;
 import org.jooq.impl.SQLDataType;
 
@@ -15,7 +13,7 @@ public class DbFieldDescriptor<DbValueT> {
     /**
      * How numerics are stored.
      */
-    public static final DataType<BigDecimal> NUMERIC_DB_TYPE = SQLDataType.DECIMAL(15, 3);
+    public static final DataType<Double> NUMERIC_DB_TYPE = SQLDataType.DOUBLE;
     /**
      * Truncated string length - backend does not store infinite strings.
      */
